# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Rest API job finder server.",
        "title": "Job Finder App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "David Afdal Kaizar Mutahadi",
            "url": "https://davidafdal.github.io/web-portofolio/",
            "email": "davidafdal7@gmail.com"
        },
        "version": "1.0"
    },
    "host": "103.157.26.216:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/job-applicants/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Apply for a job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job-applicants"
                ],
                "summary": "Apply for a job",
                "parameters": [
                    {
                        "description": "Apply Job Request",
                        "name": "applyJob",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/binder.ApplyJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/job-applicants/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approve job applicant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job-applicants"
                ],
                "summary": "Approve job applicant",
                "parameters": [
                    {
                        "description": "Approve Applicant Request",
                        "name": "approveApplicant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/binder.ApproveApplicantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/job-applicants/withdraw": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Withdraw job application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job-applicants"
                ],
                "summary": "Withdraw job application",
                "parameters": [
                    {
                        "description": "Withdraw Job Request",
                        "name": "withdrawJob",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/binder.WithdrawJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/job-applicants/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Find job applicant by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job-applicants"
                ],
                "summary": "Find job applicant by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job Applicant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "binder.ApplyJobRequest": {
            "type": "object",
            "properties": {
                "jobID": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "binder.ApproveApplicantRequest": {
            "type": "object",
            "properties": {
                "jobApplicantID": {
                    "type": "string"
                }
            }
        },
...
        "binder.WithdrawJobRequest": {
            "type": "object",
            "properties": {
                "jobApplicantID": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}